requirements.txt
<<<<<<< HEAD
streamlit==1.34.0
=======
streamlit>=1.28.0
>>>>>>> d6449752cb763615cf7462faa53dd6016407ce3d
google-generativeai
python-dotenv
pathlib
yt-dlp
requests
torch
whisper
youtube-transcript-api==1.2.2
